\docType{class}
\name{LMOutput-class}
\alias{LMOutput-class}
\alias{getcoef,LMOutput-method}
\alias{getrSq,LMOutput-method}
\alias{initialize,runLM-method}
\title{A linear model output object}
\description{
Objects of class \code{LMOutput} are created by the
\code{runLM} function Print function displays output Plot
function plots the coefficients one at a time, displaying
their values for each test. Show function runs the print
and plot functions
}
\details{
An object of the class `LMOutput' has the following slots:
\itemize{ \item \code{coef} Matrix of estimated
coefficients of the linear models \item \code{rSq} A vector
of R-squared values for the linear models }
}
\author{
David G. Carlson \email{carlson.david@wustl.edu}
}

